requirements.txt
Pillow==7.2.0
Flask==1.1.1
gunicorn==20.0.4
